New entry in database, returns string path planning could not make a path, minor optimisations
New entry added into the database for a point that is unreachable from any other point. To simulate any potential island situations that may arise in deployment of the product. Optimisations applied so that forwards the option to exit early if the function has been completed. Path planning was change the it now returns an output if it is not able to create a path between points. Should be noted it will try at all costs to create a path and will search through the entire map environment and even loading new tables if necessary